Changing schema for exact storage of wattages from hobo

+++ database/entrypoint/create_db.sql
@@ -12,11 +12,11 @@ CREATE TABLE "log" (
 	"id" SERIAL PRIMARY KEY,
 	"building" CHAR(4) NOT NULL,
 	"datetime" TIMESTAMP NOT NULL,
-	"kitchen" REAL DEFAULT 0,
-	"plugs" REAL DEFAULT 0,
-	"lights" REAL DEFAULT 0,
-	"solar" REAL DEFAULT 0,
-	"ev" REAL DEFAULT 0
+	"kitchen" numeric(8,2) DEFAULT 0,
+	"plugs" numeric(8,2) DEFAULT 0,
+	"lights" numeric(8,2) DEFAULT 0,
+	"solar" numeric(8,2) DEFAULT 0,
+	"ev" numeric(8,2) DEFAULT 0
 );
